pyproject.toml
[tool.poetry]
name = "ramp-aging-report"
version = "0.1.0"
description = ""
authors = ["Ahmet Besiroglu <ahmetybesiroglu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
